# chaos-api/src/main/resources/static/doc/openapi.json
{
  "components": {
    "schemas": {
...
      "CommonResultSysOperationLogVO_response": {
        "properties": {
          "success": {
            "format": "boolean",
            "description": "是否成功",
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "format": "string",
            "description": "响应信息",
            "type": "string",
            "example": ""
          },
          "data": {
            "description": "响应数据(object)",
            "type": "object",
            "$ref": "#/components/schemas/SysOperationLogVO_response"
          },
          "code": {
            "format": "string",
            "description": "响应码",
            "type": "string",
            "example": ""
          },
          "timestamp": {
            "format": "string",
            "description": "时间戳",
            "type": "string",
            "example": ""
          },
          "traceId": {
            "format": "string",
            "description": "请求ID",
            "type": "string",
            "example": ""
          }
        }
      },
...
      "SysOperationLogVO_response": {
        "properties": {
          "id": {
            "format": "int64",
            "description": "日志主键",
            "type": "number",
            "example": "0"
          },
          "operationIp": {
            "format": "string",
            "description": "IP地址",
            "type": "string",
            "example": ""
          },
          "operationType": {
            "format": "int32",
            "description": "操作类型(0其它 1新增 2修改 3删除)",
            "type": "integer",
            "example": "0"
          },
          "operationMethod": {
            "format": "string",
            "description": "方法名",
            "type": "string",
            "example": ""
          },
          "requestParams": {
            "format": "string",
            "description": "请求参数",
            "type": "string",
            "example": ""
          },
          "responseContent": {
            "format": "string",
            "description": "响应内容",
            "type": "string",
            "example": ""
          },
          "description": {
            "format": "string",
            "description": "操作内容描述",
            "type": "string",
            "example": ""
          },
          "operationName": {
            "format": "string",
            "description": "操作用户名",
            "type": "string",
            "example": ""
          },
          "operationTime": {
            "format": "string",
            "description": "操作时间",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "duration": {
            "format": "int64",
            "description": "消耗时间",
            "type": "number",
            "example": "0"
          }
        }
      },
      "CommonResultSysDeptVO_response": {
        "properties": {
          "success": {
            "format": "boolean",
            "description": "是否成功",
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "format": "string",
            "description": "响应信息",
            "type": "string",
            "example": ""
          },
          "data": {
            "description": "响应数据(object)",
            "type": "object",
            "$ref": "#/components/schemas/SysDeptVO_response"
          },
          "code": {
            "format": "string",
            "description": "响应码",
            "type": "string",
            "example": ""
          },
          "timestamp": {
            "format": "string",
            "description": "时间戳",
            "type": "string",
            "example": ""
          },
          "traceId": {
            "format": "string",
            "description": "请求ID",
            "type": "string",
            "example": ""
          }
        }
      },
      "SysOperationLogUpDTO_request": {
        "properties": {
          "id": {
            "format": "int64",
            "description": "日志主键",
            "type": "number",
            "example": "0"
          },
          "operationIp": {
            "format": "string",
            "description": "IP地址",
            "type": "string"
          },
          "operationType": {
            "format": "int32",
            "description": "操作类型(0其它 1新增 2修改 3删除)",
            "type": "integer",
            "example": "0"
          },
          "operationMethod": {
            "format": "string",
            "description": "方法名",
            "type": "string"
          },
          "requestParams": {
            "format": "string",
            "description": "请求参数",
            "type": "string"
          },
          "responseContent": {
            "format": "string",
            "description": "响应内容",
            "type": "string"
          },
          "description": {
            "format": "string",
            "description": "操作内容描述",
            "type": "string"
          },
          "operationName": {
            "format": "string",
            "description": "操作用户名",
            "type": "string"
          },
          "operationTime": {
            "format": "string",
            "description": "操作时间",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "duration": {
            "format": "int64",
            "description": "消耗时间",
            "type": "number",
            "example": "0"
          }
        }
      },
...
      "SysDeptUpDTO_request": {
        "properties": {
          "id": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "parentId": {
            "format": "int64",
            "description": "父部门id",
            "type": "number",
            "example": "0"
          },
          "ancestors": {
            "format": "string",
            "description": "祖级列表",
            "type": "string"
          },
          "deptName": {
            "format": "string",
            "description": "部门名称",
            "type": "string"
          },
          "orderNum": {
            "format": "int32",
            "description": "显示顺序",
            "type": "integer",
            "example": "0"
          }
        }
      },
      "SysDeptVO_response": {
        "properties": {
          "id": {
            "format": "int64",
            "description": "部门id",
            "type": "number",
            "example": "0"
          },
          "parentId": {
            "format": "int64",
            "description": "父部门id",
            "type": "number",
            "example": "0"
          },
          "ancestors": {
            "format": "string",
            "description": "祖级列表",
            "type": "string",
            "example": ""
          },
          "deptName": {
            "format": "string",
            "description": "部门名称",
            "type": "string",
            "example": ""
          },
          "orderNum": {
            "format": "int32",
            "description": "显示顺序",
            "type": "integer",
            "example": "0"
          }
        }
      },
...
      "CommonResultBoolean_response": {
        "properties": {
          "success": {
            "format": "boolean",
            "description": "是否成功",
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "format": "string",
            "description": "响应信息",
            "type": "string",
            "example": ""
          },
          "data": {
            "format": "boolean",
            "description": "响应数据",
            "type": "boolean"
          },
          "code": {
            "format": "string",
            "description": "响应码",
            "type": "string",
            "example": ""
          },
          "timestamp": {
            "format": "string",
            "description": "时间戳",
            "type": "string",
            "example": ""
          },
          "traceId": {
            "format": "string",
            "description": "请求ID",
            "type": "string",
            "example": ""
          }
        }
      },
      "SysOperationLogDTO_request": {
        "properties": {
          "operationIp": {
            "format": "string",
            "description": "IP地址",
            "type": "string"
          },
          "operationType": {
            "format": "int32",
            "description": "操作类型(0其它 1新增 2修改 3删除)",
            "type": "integer",
            "example": "0"
          },
          "operationMethod": {
            "format": "string",
            "description": "方法名",
            "type": "string"
          },
          "requestParams": {
            "format": "string",
            "description": "请求参数",
            "type": "string"
          },
          "responseContent": {
            "format": "string",
            "description": "响应内容",
            "type": "string"
          },
          "description": {
            "format": "string",
            "description": "操作内容描述",
            "type": "string"
          },
          "operationName": {
            "format": "string",
            "description": "操作用户名",
            "type": "string"
          },
          "operationTime": {
            "format": "string",
            "description": "操作时间",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "duration": {
            "format": "int64",
            "description": "消耗时间",
            "type": "number",
            "example": "0"
          }
        }
      },
...
      "SysDeptDTO_request": {
        "properties": {
          "parentId": {
            "format": "int64",
            "description": "父部门id",
            "type": "number",
            "example": "0"
          },
          "ancestors": {
            "format": "string",
            "description": "祖级列表",
            "type": "string"
          },
          "deptName": {
            "format": "string",
            "description": "部门名称",
            "type": "string"
          },
          "orderNum": {
            "format": "int32",
            "description": "显示顺序",
            "type": "integer",
            "example": "0"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "openapi": "3.0.3",
  "paths": {
    "/api/v1/sysOperationLog/edit": {
      "put": {
        "summary": "更新操作日志信息",
        "deprecated": false,
        "description": "更新操作日志信息",
        "tags": [
          "基础模块-操作日志"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysOperationLogUpDTO_request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": 0,\n  \"operationIp\": \"\",\n  \"operationType\": 0,\n  \"operationMethod\": \"\",\n  \"requestParams\": \"\",\n  \"responseContent\": \"\",\n  \"description\": \"\",\n  \"operationName\": \"\",\n  \"operationTime\": \"yyyy-MM-dd HH:mm:ss\",\n  \"duration\": 0\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": true,\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-sysOperationLog-edit-PUT",
        "parameters": []
      }
    },
...
    "/api/v1/sysOperationLog/save": {
      "post": {
        "summary": "保存操作日志信息",
        "deprecated": false,
        "description": "保存操作日志信息",
        "tags": [
          "基础模块-操作日志"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysOperationLogDTO_request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"operationIp\": \"\",\n  \"operationType\": 0,\n  \"operationMethod\": \"\",\n  \"requestParams\": \"\",\n  \"responseContent\": \"\",\n  \"description\": \"\",\n  \"operationName\": \"\",\n  \"operationTime\": \"yyyy-MM-dd HH:mm:ss\",\n  \"duration\": 0\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": true,\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-sysOperationLog-save-POST",
        "parameters": []
      }
    },
...
    "/api/v1/dept/{ids}": {
      "delete": {
        "summary": "删除部门信息",
        "deprecated": false,
        "description": "删除部门信息",
        "tags": [
          "基础模块-部门管理"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": true,\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-dept-{ids}-DELETE",
        "parameters": [
          {
            "schema": {
              "type": "object"
            },
            "in": "path",
            "description": "部门id数组,[array of int64]",
            "required": true,
            "name": "ids"
          }
        ]
      }
    },
    "/api/v1/dept/save": {
      "post": {
        "summary": "保存部门信息",
        "deprecated": false,
        "description": "保存部门信息",
        "tags": [
          "基础模块-部门管理"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysDeptDTO_request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"parentId\": 0,\n  \"ancestors\": \"\",\n  \"deptName\": \"\",\n  \"orderNum\": 0\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": true,\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-dept-save-POST",
        "parameters": []
      }
    },
...
    "/api/v1/sysOperationLog/{ids}": {
      "delete": {
        "summary": "删除操作日志信息",
        "deprecated": false,
        "description": "删除操作日志信息",
        "tags": [
          "基础模块-操作日志"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": true,\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-sysOperationLog-{ids}-DELETE",
        "parameters": [
          {
            "schema": {
              "type": "object"
            },
            "in": "path",
            "description": "操作日志id数组,[array of int64]",
            "required": true,
            "name": "ids"
          }
        ]
      }
    },
    "/api/v1/sysOperationLog/{id}": {
      "get": {
        "summary": "根据id查询操作日志信息",
        "deprecated": false,
        "description": "根据id查询操作日志信息",
        "tags": [
          "基础模块-操作日志"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultSysOperationLogVO_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": {\n    \"id\": 0,\n    \"operationIp\": \"\",\n    \"operationType\": 0,\n    \"operationMethod\": \"\",\n    \"requestParams\": \"\",\n    \"responseContent\": \"\",\n    \"description\": \"\",\n    \"operationName\": \"\",\n    \"operationTime\": \"yyyy-MM-dd HH:mm:ss\",\n    \"duration\": 0\n  },\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-sysOperationLog-{id}-GET",
        "parameters": [
          {
            "schema": {
              "format": "integer",
              "type": "int64"
            },
            "in": "path",
            "description": "操作日志id",
            "required": true,
            "name": "id"
          }
        ]
      }
    },
...
    "/api/v1/dept/edit": {
      "put": {
        "summary": "更新部门信息",
        "deprecated": false,
        "description": "更新部门信息",
        "tags": [
          "基础模块-部门管理"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysDeptUpDTO_request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": 0,\n  \"parentId\": 0,\n  \"ancestors\": \"\",\n  \"deptName\": \"\",\n  \"orderNum\": 0\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": true,\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-dept-edit-PUT",
        "parameters": []
      }
    },
...
    "/api/v1/dept/{id}": {
      "get": {
        "summary": "根据id查询部门信息",
        "deprecated": false,
        "description": "根据id查询部门信息",
        "tags": [
          "基础模块-部门管理"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultSysDeptVO_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": {\n    \"id\": 0,\n    \"parentId\": 0,\n    \"ancestors\": \"\",\n    \"deptName\": \"\",\n    \"orderNum\": 0\n  },\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-dept-{id}-GET",
        "parameters": [
          {
            "schema": {
              "format": "integer",
              "type": "int64"
            },
            "in": "path",
            "description": "部门id",
            "required": true,
            "name": "id"
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "chaos接口文档",
    "version": "1.0.0"
  },
...
}